refactor: extract background-process/open-url.js
